Continued Happiness and Wellbeing balance; Happiness bar has been replaced with states; Rebalanced nodes for happiness; Nodes now have a unit per tile limit; Additional intro map updates; Commands for refilling and auto-refill now automatically generate; Camera moves smoother; Improved movement for projectiles so they now use lerping; Hobbler battle positions are now saved in data; Fixed a bug where hobblers could work and battle at the same time; Hobblers now have an experience pool that is used for Studying; Studying now distributes the monster's experience to other monsters; Training Wand now gives mana for each use to help balance starting casters; Nodes set to auto-refill now save their state; Players now lose gold when being defeated in battle; Dialogue can now apply traits to the owner and player once finished; Additional improvements to Summerset Hollow

+++ Assets/Ancible Tools/Scripts/Traits/MinigameCombatStatsTrait.cs
@@ -21,7 +21,7 @@ namespace Assets.Ancible_Tools.Scripts.Traits
         private CombatStats _bonusStats = CombatStats.Zero;
 
         private int _currentHealth = 0;
-        private int _currentMana = 0;
+        private float _currentMana = 0;
 
         public override void SetupController(TraitController controller)
         {
@@ -169,7 +169,7 @@ namespace Assets.Ancible_Tools.Scripts.Traits
 
         private void QueryMana(QueryManaMessage msg)
         {
-            msg.DoAfter.Invoke(_currentMana, _baseStats.Mana + _bonusStats.Mana);
+            msg.DoAfter.Invoke((int)_currentMana, _baseStats.Mana + _bonusStats.Mana);
         }
 
         private void Heal(HealMessage msg)
